[refactor] #103 알림창 아이콘 png -> 벡터이미지 변경
