+++ background/src/main/java/com/dpm/sixpack/runningservice/di/ServiceModule.kt
@@ -23,7 +23,7 @@ object ServiceModule {
         .Builder(context, NOTIFICATION_CHANNEL_ID)
         .setAutoCancel(false)
         .setOngoing(true)
-        .setSmallIcon(R.drawable.img_notification_icon)
+        .setSmallIcon(R.drawable.ic_notification)
         .setContentTitle("달리기 기록을 측정중입니다.")
         .setContentText("00:00:00")
 //        .setContentIntent(pendingIntent)
