+++ background/src/main/res/drawable/ic_notification.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="83dp"
+    android:height="78dp"
+    android:viewportWidth="83"
+    android:viewportHeight="78">
+  <path
+      android:pathData="M42.033,0C54.421,0 63.334,2.19 68.772,6.571C74.286,10.877 77.043,16.731 77.043,24.133C77.043,28.589 75.834,32.592 73.417,36.143C71.529,38.856 68.742,41.223 65.057,43.245C63.696,43.992 63.066,45.853 64.327,46.759C65.8,47.816 68.642,48.462 70.695,48.586C71.742,48.649 73.334,48.688 74.644,48.711H78.476C80.431,48.711 82.016,50.296 82.016,52.251V76.704C82.016,77.372 81.474,77.914 80.806,77.914C75.98,77.914 71.814,75.09 69.871,71.005C69.212,69.62 67.878,68.563 66.345,68.563C62.53,68.563 56.412,67.809 52.862,66.185C49.312,64.561 46.479,62.256 44.364,59.386C42.25,56.516 40.66,53.142 39.606,49.303C39.405,48.569 38.756,48.04 37.995,48.04H37.291C35.905,48.04 34.781,49.164 34.781,50.55V74.41C34.781,76.366 33.196,77.951 31.241,77.951H27.034C27.034,77.951 27.033,77.952 27.033,77.952C27.033,77.953 27.033,77.953 27.032,77.953H1.21C0.542,77.953 0,77.411 0,76.743C0,71.827 2.931,67.596 7.141,65.701C8.706,64.997 9.969,63.578 9.969,61.862V14.519C9.969,6.501 16.469,0 24.488,0H42.033ZM34.781,28.977C34.781,30.932 36.367,32.518 38.322,32.518H39.087C44.676,32.518 48.604,31.913 50.87,30.704C53.212,29.42 54.383,27.267 54.383,24.246C54.383,21.678 53.438,19.827 51.55,18.694C49.661,17.486 46.678,16.882 42.6,16.882H38.322C36.367,16.882 34.781,18.467 34.781,20.423V28.977Z"
+      android:fillColor="#ffffff"/>
+</vector>
