+++ background/src/main/res/drawable/ill_logo.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="200dp"
+    android:height="200dp"
+    android:viewportWidth="200"
+    android:viewportHeight="200">
+  <path
+      android:pathData="M24,0L176,0A24,24 0,0 1,200 24L200,176A24,24 0,0 1,176 200L24,200A24,24 0,0 1,0 176L0,24A24,24 0,0 1,24 0z"
+      android:fillColor="#3E4FFF"/>
+  <path
+      android:pathData="M63.94,61.02C76.33,61.02 85.24,63.21 90.68,67.59C96.2,71.9 98.95,77.75 98.95,85.16C98.95,89.61 97.74,93.62 95.33,97.17C93.44,99.88 90.65,102.25 86.96,104.27C85.6,105.02 84.97,106.88 86.23,107.78C87.71,108.84 90.55,109.48 92.6,109.61C93.69,109.67 95.36,109.71 96.7,109.73H100.38C102.15,109.73 103.61,111.03 103.88,112.72V122.86C103.88,123.63 103.28,124.53 102.12,125.05C97.98,126.91 94.91,130.74 94.11,135.35C93.16,134.39 92.37,133.26 91.78,132.03C91.12,130.64 89.79,129.59 88.25,129.59C84.44,129.59 78.32,128.83 74.77,127.21C71.22,125.58 68.39,123.28 66.27,120.41C64.16,117.54 62.57,114.16 61.51,110.32C61.31,109.59 60.66,109.06 59.91,109.06H59.2C57.81,109.06 56.69,110.19 56.69,111.57V135.43C56.69,137.39 55.11,138.98 53.15,138.98H48.94L48.94,138.98L48.94,138.98H23.12C22.45,138.98 21.91,138.44 21.91,137.77C21.91,132.85 24.84,128.62 29.05,126.72C30.62,126.02 31.88,124.6 31.88,122.88V75.54C31.88,67.52 38.38,61.02 46.4,61.02H63.94ZM60.23,77.91C58.28,77.91 56.69,79.49 56.69,81.45V90C56.69,91.96 58.28,93.54 60.23,93.54H61C66.59,93.54 70.51,92.94 72.78,91.73C75.12,90.44 76.29,88.29 76.29,85.27C76.29,82.7 75.35,80.85 73.46,79.72C71.57,78.51 68.59,77.91 64.51,77.91H60.23Z"
+      android:fillColor="#ffffff"/>
+  <path
+      android:pathData="M138.1,61.02C150.49,61.02 159.4,63.21 164.84,67.59C170.35,71.9 173.11,77.75 173.11,85.16C173.11,89.61 171.9,93.62 169.48,97.17C167.6,99.88 164.81,102.25 161.12,104.27C159.76,105.02 159.13,106.88 160.39,107.78C161.86,108.84 164.71,109.48 166.76,109.61C167.85,109.67 169.52,109.71 170.86,109.73H174.54C176.49,109.73 178.08,111.32 178.08,113.28V137.73C178.08,138.4 177.54,138.94 176.87,138.94C172.04,138.94 167.88,136.11 165.93,132.03C165.28,130.64 163.94,129.59 162.41,129.59C158.59,129.59 152.48,128.83 148.93,127.21C145.38,125.58 142.54,123.28 140.43,120.41C138.31,117.54 136.72,114.16 135.67,110.32C135.47,109.59 134.82,109.06 134.06,109.06H133.36C131.97,109.06 130.85,110.19 130.85,111.57V135.43C130.85,137.39 129.26,138.98 127.31,138.98H123.1C123.1,138.98 123.1,138.98 123.1,138.98C123.1,138.98 123.1,138.98 123.09,138.98H97.27C96.6,138.98 96.06,138.44 96.06,137.77C96.06,132.85 98.99,128.62 103.21,126.72C104.77,126.02 106.04,124.6 106.04,122.88V75.54C106.04,67.52 112.54,61.02 120.56,61.02H138.1ZM130.85,90C130.85,91.96 132.43,93.54 134.39,93.54H135.15C140.74,93.54 144.67,92.94 146.94,91.73C149.28,90.44 150.45,88.29 150.45,85.27C150.45,82.7 149.51,80.85 147.62,79.72C145.73,78.51 142.74,77.91 138.67,77.91H134.39C132.43,77.91 130.85,79.49 130.85,81.45V90Z"
+      android:fillColor="#ffffff"/>
+</vector>
